Rename: change the BaseTimeEntity package
